modulo02-python-com-banco-de-dados/20240203-aula04/requirements.txt
greenlet==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
